Encapsulate legend rendering code
